image_classification_app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
bottleneck=1.3.2=py39hdd57654_1
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py39hf3d152e_1
cffi=1.14.6=py39h400218f_0
click=8.0.3=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
flask=2.0.2=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
intel-openmp=2019.4=243
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libmklml=2019.0.5=0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py39h27cfd23_0
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.3.0=py39h54f3939_0
mkl_random=1.0.2=py39h63df603_0
ncurses=6.3=h7f8727e_0
ninja=1.10.2=hff7bd54_1
numexpr=2.7.3=py39hb2eb853_0
numpy=1.19.2=py39h89c1606_0
numpy-base=1.19.2=py39h2ae0177_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1l=h7f8727e_0
pandas=1.3.4=py39h8c16a72_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
pycparser=2.20=py_2
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.19.1=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.7.1=cpu_py39h6a09485_0
pytz=2021.3=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.11=h1ccaba5_0
torchvision=0.8.2=cpu_py39ha229d99_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
